Adds "devcap2subckt" option to ext2spice
When active, the option transforms the device capacitors
from a "C" card to a "X" card (a subckt instance) and adds
an additional connection to the substrate node

e.g.

CN a b capmodel l=2.0u w=2.0u

becomes

XM a b subnode capmodel l=2.0u w=2.0u

The instance numbers "N" and "M" are whatever number is next
for that type of instance.

Two significant changes:

- char *substr   is made global so that it can be availabe in the
EFVisitDevs function.

- call to EFVisitNodes moved to the beginning of the routine
so that the substrate node name is found before calling
EFVisitDevs

This only affects device capacitors which have been extracted.
Future enhancement would be to designate which particular
devices it affects, and whether or not to add the substrate node.