Important update:  Reworked the extraction method to properly isolate the terminal areas of a device (e.g., source and drain) and calculate their area and perimeter individually for the device (in addition to the traditional method of calculating area and perimeter of each resistance class for the entire node).
Also:  Reworked the SPICE syntax output to generate SI values
in the range 1-1000 with the appropriate suffix (e.g., "20u")
instead of defaulting to "u" for lengths and "p" for areas.
This prevents it from producing weird units like "150000u" when
a process definition already includes a scalefactor.

Reworked the "extresist" code to use the device terminal area
and perimeter.  This fixes an error in which "extresist" would
lose these values and "ext2spice" with option "extresist on"
would generate a new netlist output with zero terminal areas
and perimeters.